+++ src/core/aio/aio_task.cpp
@@ -2,7 +2,7 @@
 // This source code is licensed under the Apache License Version 2.0, which
 // can be found in the LICENSE file in the root directory of this source tree.
 
-#include "core/concurrent/task_engine.h"
+#include "core/task/task_engine.h"
 #include <dsn/tool-api/file_io.h>
 #include <dsn/utility/error_code.h>
