+++ src/core/core/service_api_c.cpp
@@ -38,7 +38,7 @@
 
 #include "service_engine.h"
 #include "core/rpc/rpc_engine.h"
-#include "core/concurrent/task_engine.h"
+#include "core/task/task_engine.h"
 #include "coredump.h"
 
 //
